Added the ability to unpack mes_all.bin from Ace Attorney for the Nintendo DS platform.
Added the ability to unpack mes_all.bin from Ace Attorney for the Nintendo DS platform. The new function is available in the menu on the path: File -> Open -> DS. So far, the function can only decompress. The unpacked scripts will appear in the same folder as the program. Please don't be lazy to drag the result to a separate folder.